Combined Navbar into one component, added Brand component, made Brand link to root directory
